requirements.txt
Flask
python-dotenv
pandas
requests
langchain
openai
langchain-openai
sentence_transformers
rapidfuzz
annoy
langgraph
langsmith
faiss-cpu
graphviz